+++ test-support/unsafe-test-value.d.ts
@@ -1,0 +1,2 @@
+export declare function unsafeTestValue<TValue extends PropertyKey>(value: TValue): TValue;
+export declare function unsafeTestValue<TValue>(value: unknown): TValue;